+++ packages/importers/src/index.ts
@@ -5,3 +5,4 @@ export * from './adapters.ts';
 export * from './association.ts';
 export * from './importer.ts';
 export * from './inventory.ts';
+export * from './folder.ts';